style: move SidebarEntityAbout
