+++ wondrous-app/components/Common/SidebarEntityAbout/SettingsButton.tsx
