+++ wondrous-app/components/Common/SidebarEntityAbout/index.tsx
